OSBot Files
OSBot was the old client I used and developed with over the summer. I have since switched to TRiBot because it is a better platform for not getting banned as much, and has many more built in features such as human-realistic mouse movements and random tendencies not available in OSBot.